[feature/article-comment] refactor : 문제 커밋 해결 (회원 수정 페이지) 이동 에러 해결
