+++ src/main/java/dev/linkcentral/presentation/controller/view/MemberViewController.java
@@ -1,25 +1,13 @@
 package dev.linkcentral.presentation.controller.view;
 
 import dev.linkcentral.database.entity.Member;
-import dev.linkcentral.infrastructure.jwt.JwtAuthenticationFilter;
-import dev.linkcentral.infrastructure.jwt.JwtTokenProvider;
 import dev.linkcentral.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.view.RedirectView;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.Collections;
 
 @Controller
 @Slf4j
@@ -50,14 +38,12 @@ public class MemberViewController {
     }
 
     @GetMapping("/edit-form")
-    public String memberEdit(Model model) {
+    public String memberEditForm(Model model) {
         Member member = memberService.getCurrentMember();
+        if (member == null) {
+            return "redirect:/login";
+        }
         model.addAttribute("member", member);
-        return "/members/edit";
-    }
-
-    @GetMapping("/edit-form2")
-    public String memberEdit2() {
         return "/members/edit";
     }
