+++ src/main/webapp/WEB-INF/views/members/login.jsp
@@ -51,7 +51,6 @@
         }
     </style>
 
-
     <script>
         function logout() {
             // JWT 토큰 삭제
@@ -67,10 +66,98 @@
                 headers: {
                     'Authorization': 'Bearer ' + localStorage.getItem("jwt")
                 },
-                success: function(data) {
-                    // window.location.href = data
+                success: function (htmlContent) {
+                    // 서버로부터 받은 HTML 콘텐츠를 현재 페이지의 DOM에 삽입
+                    document.body.innerHTML = htmlContent; // 현재 페이지 바디에 폼 삽입
+
+                    // 시작점
+                    $(document).ready(function () {
+                        init();
+
+                        function init() {
+                            $("#btn-update").on("click", function () {
+                                checkCurrentPassword();
+                            });
+                        }
+
+                        function checkCurrentPassword() {
+                            let currentPassword = $("#currentPassword").val();
+
+                            $(".error-message").remove();
+
+                            $.ajax({
+                                url: "/api/v1/public/member/check-current-password",
+                                type: "POST",
+                                data: { password: currentPassword }, // Data is sent as form parameters
+                                headers: {
+                                    'Authorization': 'Bearer ' + localStorage.getItem("jwt")
+                                },
+                                success: function (resp) {
+                                    console.log("Response:", resp);
+
+                                    if (resp.result) {
+                                        update();
+                                    } else {
+                                        let errorMessage = "현재 비밀번호가 일치하지 않거나, 입력하지 않는 문구가 있습니다.";
+                                        $("#currentPassword").after("<div class='error-message'>" + errorMessage + "</div>");
+                                    }
+                                },
+                                error: function (xhr, status, error) {
+                                    console.error("Error checking current password:", xhr.responseText);
+                                    let errorMessage = "An error occurred. Please try again.";
+                                    $("#currentPassword").after("<div class='error-message'>" + errorMessage + "</div>");
+                                }
+                            });
+                        }
+
+                        function update() {
+                            let data = {
+                                id: $("#id").val(),
+                                name: $("#name").val(),
+                                password: $("#password").val(),
+                                nickname: $("#nickname").val()
+                            };
+
+                            console.log("업데이트 데이터:", data);
+
+                            // 데이터 객체를 폼 데이터로 변환
+                            let formData = new FormData();
+
+                            // 빈 값 또는 null 값이 아닌 경우에만 FormData에 추가
+                            Object.keys(data).forEach(key => {
+                                if (data[key] !== null && data[key] !== "") {
+                                    formData.append(key, data[key]);
+                                }
+                            });
+
+                            $.ajax({
+                                type: "PUT",
+                                url: "/api/v1/member/edit",
+                                data: formData,
+                                headers: {
+                                    'Authorization': 'Bearer ' + localStorage.getItem("jwt")
+                                },
+                                processData: false,
+                                contentType: false
+                            })
+                                .done(function (resp) {
+                                    alert("회원 수정이 완료되었습니다.");
+                                    window.location.href = "/";
+                                })
+                                .fail(function (error) {
+                                    let errorMessage = "현재 비밀번호가 일치하지 않습니다.";
+
+                                    if ($("#currentPassword + .error-message").length === 0) {
+                                        $("#currentPassword").after("<div class='error-message'>" + errorMessage + "</div>");
+                                    }
+                                });
+                        }
+                    });
                 },
-                error: function(error) {
+                error: function (error) {
+                    console.error(error);
+                    // 에러 처리, 예를 들어 사용자에게 알림을 보여주는 등
+                    alert('프로필 정보를 가져오는데 실패했습니다.');
                 }
             });
         }
@@ -90,7 +177,7 @@
                 headers: {
                     'Authorization': 'Bearer ' + localStorage.getItem("jwt")
                 },
-                success: function(response) {
+                success: function (response) {
                     var memberId = response.memberId;
                     if (memberId === null || memberId === '') {
                         alert('회원 ID가 존재하지 않습니다.');
@@ -98,7 +185,7 @@
                     }
                     window.location.href = "/api/v1/view/profile/view?memberId=" + memberId;
                 },
-                error: function(xhr) {
+                error: function (xhr) {
                     alert("회원 정보를 가져올 수 없습니다: " + xhr.responseText);
                 }
             });
